fixed mel extraction errors

+++ audioldm2/utilities/audio/tools.py
@@ -40,7 +40,7 @@ def read_wav_file(filename, segment_length):
     return waveform
 
 def get_mel_from_wav(audio, _stft):
-    audio = torch.clip(torch.FloatTensor(audio), -1, 1)
+    audio = torch.clip(audio, -1, 1)
     audio = torch.autograd.Variable(audio, requires_grad=False)
     melspec, magnitudes, phases, energy = _stft.mel_spectrogram(audio)
     melspec = torch.squeeze(melspec, 0)
